+++ src/main/java/game/Main.java
@@ -1,5 +1,7 @@
 package game;
 
+import game.domain.RacingGame;
+
 public class Main {
 
     public static void main(String[] args) {
