+++ src/test/java/game/domain/CarTest.java
@@ -1,6 +1,6 @@
-package game;
+package game.domain;
 
-import org.assertj.core.api.Assertions;
+import game.domain.Car;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
