+++ src/test/java/game/domain/RacingGameTest.java
@@ -1,5 +1,6 @@
-package game;
+package game.domain;
 
+import game.domain.RacingGame;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
